[core]: work on create function to create custom metadata - create methods setMetadata and getMetadata for custom metadata reflection - rename AbyssApplication to AbyssalApplication - rename the import for express-runner
