+++ packages/@core/src/utils/metadataUtils/index.ts
@@ -1,0 +1,4 @@
+export * from './controller';
+export * from './core';
+export * from './injection';
+export * from './ioc';
